# Backend/docs/swagger.yaml
definitions:
...
  controllers.GuruQRResponse:
    properties:
      message:
        example: QR Code berhasil dibuat, berlaku 30 detik!
        type: string
      qr_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        type: string
    type: object
  controllers.GuruStatsResponse:
    properties:
      absent_today:
        example: 2
        type: integer
      attendance_rate:
        example: 93.33
        type: number
      present_today:
        example: 28
        type: integer
      total_students:
        example: 30
        type: integer
    type: object
  controllers.JadwalGuru:
    properties:
      id:
        example: 1
        type: integer
      kelas:
        example: XII RPL 1
        type: string
      mata_pelajaran:
        example: Pemrograman Web
        type: string
      waktu_mulai:
        example: "08:00"
        type: string
      waktu_selesai:
        example: "10:00"
        type: string
    type: object
...
  controllers.ResetPasswordReq:
    properties:
      new_password:
        example: PasswordBaru123!
        type: string
      token:
        example: 817cdc18e67140e57f62dd6c80fd52529...
        type: string
    type: object
...
  controllers.StudentStatusList:
    properties:
      nama:
        example: Hafizh Junior
        type: string
      nis:
        example: "22010101"
        type: string
      status:
        example: hadir
        type: string
      waktu_absen:
        example: "2026-04-08 09:15:00"
        type: string
    type: object
  controllers.SuccessMessage:
    properties:
      message:
        example: Aksi berhasil dilakukan!
        type: string
    type: object
info:
  contact:
    email: hafizh@example.com
    name: Hafizh
  description: Ini adalah dokumentasi REST API untuk aplikasi absensi dengan fitur
    QR & Geofencing.
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Tempat Les
  version: "1.0"
paths:
...
  /guru/backup:
    get:
      description: Melakukan dump SQL dan mengunggahnya otomatis ke folder Drive
      produces:
      - application/json
      responses:
        "200":
          description: Backup Berhasil Diupload
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Backup Database ke Google Drive
      tags:
      - 3. Operasional Guru
  /guru/dashboard/attendance-list:
    get:
      description: Menampilkan daftar nama siswa dan status absensinya (Hadir/Belum)
      parameters:
      - description: ID Sesi Pembelajaran
        in: query
        name: sesi_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar Kehadiran
          schema:
            items:
              $ref: '#/definitions/controllers.StudentStatusList'
            type: array
      security:
      - BearerAuth: []
      summary: Detail Status Kehadiran Per Siswa
      tags:
      - 3. Operasional Guru
  /guru/dashboard/stats:
    get:
      description: Menampilkan total murid terdaftar, hadir, dan rasio kehadiran
      parameters:
      - description: ID Sesi Pembelajaran
        in: query
        name: sesi_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Data Ditemukan
          schema:
            $ref: '#/definitions/controllers.GuruStatsResponse'
      security:
      - BearerAuth: []
      summary: Ringkasan Statistik Kelas
      tags:
      - 3. Operasional Guru
  /guru/export:
    get:
      description: Otomatis mengunduh file Excel laporan absensi dan mem-backup ke
        G-Drive
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: File Downloaded
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Download Laporan Excel
      tags:
      - 3. Operasional Guru
  /guru/generate-qr:
    get:
      description: Menghasilkan JWT token beralaku 30 detik untuk diubah jadi gambar
        QR
      parameters:
      - description: ID Sesi Pembelajaran
        in: query
        name: sesi_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses Membuat QR
          schema:
            $ref: '#/definitions/controllers.GuruQRResponse'
      security:
      - BearerAuth: []
      summary: Tampilkan QR Code (Guru)
      tags:
      - 3. Operasional Guru
  /guru/jadwal:
    get:
      description: Menampilkan sesi pembelajaran berdasarkan guru yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: Daftar Jadwal
          schema:
            items:
              $ref: '#/definitions/controllers.JadwalGuru'
            type: array
      security:
      - BearerAuth: []
      summary: Jadwal Mengajar Guru
      tags:
      - 3. Operasional Guru
  /guru/reset-password:
    post:
      consumes:
      - application/json
      description: Mengganti kata sandi guru menggunakan token rahasia
      parameters:
      - description: Token dan Password Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password berhasil diubah
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      summary: Eksekusi Reset Password
      tags:
      - 1. Auth & Keamanan
  /guru/restore:
    post:
      consumes:
      - application/json
      description: Mengunduh file SQL dari Drive menggunakan File ID dan menimpanya
        ke database lokal
      parameters:
      - description: Google Drive File ID
        in: query
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Database Berhasil Direstore
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Restore Database dari Google Drive
      tags:
      - 3. Operasional Guru
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"